requirements.txt
openai
elsaticsearch  
flask
python-dotenv